// src/lib/api/generated/model/fichaFieldType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type FichaFieldType = typeof FichaFieldType[keyof typeof FichaFieldType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FichaFieldType = {
  text: 'text',
  number: 'number',
  date: 'date',
  select: 'select',
  boolean: 'boolean',
  photo: 'photo',
} as const;